--- src/app/app.component.css
@@ -1,59 +1,0 @@
-.parent {
-  background-color: whitesmoke;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-.app-header {
-  height: 60px;
-}
-
-.page-content {
-  display: flex;
-  flex-direction: row;
-  height: calc(100vh - 60px);
-}
-
-.sidebar {
-  width: 200px;
-  background-color: #f8f9fa;
-  padding: 20px;
-  box-shadow: 2px 0 8px rgba(0, 0, 0, 0.15);
-  position: sticky;
-  top: 0;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.sidebar nav ul {
-  list-style-type: none;
-}
-
-.sidebar nav ul li a {
-  color: darkblue;
-  text-decoration: none;
-  font-weight: 500;
-  display: block;
-  padding: 10px 15px;
-  border-radius: 5px;
-  margin-bottom: 10px;
-}
-
-.sidebar nav ul li a:hover {
-  background-color: darkblue;
-  color: #fff;
-}
-
-.content {
-  flex-grow: 1;
-  padding: 20px;
-  background-color: whitesmoke;
-  overflow-y: auto;
-}
-
-.app-search,
-.app-tasks {
-  margin-bottom: 20px;
-}
