+++ src/app/app.routes.ts
@@ -1,23 +1,11 @@
-import { Routes, withRouterConfig } from '@angular/router';
-import { TasksComponent } from './tasks/tasks.component';
-import { UpcomingTasksComponent } from './upcoming-tasks/upcoming-tasks.component';
-import { CompletedTasksComponent } from './completed-tasks/completed-tasks.component';
-import { OverdueTasksComponent } from './overdue-tasks/overdue-tasks.component';
-import { TaskFormComponent } from './task-form/task-form.component';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { DashboardStatsComponent } from './dashboard-stats/dashboard-stats.component';
+import { Routes } from '@angular/router';
+import { TasksComponent } from './components/tasks/tasks.component';
+import { TaskFormComponent } from './components/tasks/components/task-form/task-form.component';
+import { PageNotFoundComponent } from './components/shared/page-not-found/page-not-found.component';
 
 export const routes: Routes = [
   { path: '', redirectTo: 'allTasks', pathMatch: 'full' },
-  {
-    path: 'dashboard',
-    component: DashboardComponent,
-    children: [
-      { path: '', redirectTo: 'stats', pathMatch: 'full' },
-      { path: 'stats', component: DashboardStatsComponent },
-    ],
-  },
+
   {
     path: 'allTasks',
     component: TasksComponent,
@@ -27,12 +15,8 @@ export const routes: Routes = [
     component: TaskFormComponent,
   },
   {
-    path: 'allTasks/:taskId/edit',
+    path: 'allTasks/edit',
     component: TaskFormComponent,
-    data: { renderMode: 'client', prerender: false, ssr: false },
   },
-  { path: 'upcoming', component: UpcomingTasksComponent },
-  { path: 'done', component: CompletedTasksComponent },
-  { path: 'overdue', component: OverdueTasksComponent },
   { path: '**', component: PageNotFoundComponent },
 ];
