--- src/app/completed-tasks/completed-tasks.component.html
@@ -1,1 +1,0 @@
-<app-tasks [tasks]="tasks"></app-tasks>
