--- src/app/completed-tasks/completed-tasks.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { CompletedTasksComponent } from './completed-tasks.component';
-
-describe('CompletedTasksComponent', () => {
-  let component: CompletedTasksComponent;
-  let fixture: ComponentFixture<CompletedTasksComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [CompletedTasksComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(CompletedTasksComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
