--- src/app/dashboard-stats/dashboard-stats.component.html
@@ -1,1 +1,0 @@
-<p>dashboard-stats works!</p>
