--- src/app/dashboard-stats/dashboard-stats.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DashboardStatsComponent } from './dashboard-stats.component';
-
-describe('DashboardStatsComponent', () => {
-  let component: DashboardStatsComponent;
-  let fixture: ComponentFixture<DashboardStatsComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [DashboardStatsComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(DashboardStatsComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
