--- src/app/dashboard-stats/dashboard-stats.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-dashboard-stats',
-  imports: [],
-  templateUrl: './dashboard-stats.component.html',
-  styleUrl: './dashboard-stats.component.css'
-})
-export class DashboardStatsComponent {
-
-}
