--- src/app/header/header.component.css
@@ -1,9 +1,0 @@
-.header {
-  background-color: darkblue;
-  color: white;
-  height: 80px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 30px;
-}
