--- src/app/header/header.component.html
@@ -1,7 +1,0 @@
-<div class="header">
-  <!-- Property binding -->
-  <img [src]="logo" />
-
-  <!-- Data binding -->
-  <p>{{ title }}</p>
-</div>
