--- src/app/header/header.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-header',
-  imports: [],
-  templateUrl: './header.component.html',
-  styleUrl: './header.component.css',
-})
-export class HeaderComponent {
-  title: string = 'TODO APP';
-  logo: string = '/src/public/favicon.ico';
-}
