--- src/app/search/search.component.html
@@ -1,18 +1,0 @@
-<!-- Event binding + Two way binding -->
-
-<div class="search-section">
-  <div class="search-container">
-    <input
-      type="text"
-      class="search-input"
-      placeholder="Search..."
-      [(ngModel)]="searchValue"
-      (input)="onInputChange($event)"
-    />
-    <button class="search-button" (click)="onSearch()">Search</button>
-  </div>
-
-  <div class="displayed-text">
-    <p>You searched for: {{ displayedText }}</p>
-  </div>
-</div>
