+++ src/app/services/toast.service.spec.ts
@@ -1,5 +1,4 @@
 import { TestBed } from '@angular/core/testing';
-
 import { ToastService } from './toast.service';
 
 describe('ToastService', () => {
