--- src/app/upcoming-tasks/upcoming-tasks.component.html
@@ -1,1 +1,0 @@
-<app-tasks [tasks]="tasks"></app-tasks>
